+++ FirebaseStorage/Sources/StorageMetadata.swift
@@ -103,11 +103,6 @@ import Foundation
   @objc public let updated: Date?
 
   /**
-   * Never used API
-   */
-  @available(*, deprecated) @objc public let storageReference: StorageReference? = nil
-
-  /**
    * Creates a Dictionary from the contents of the metadata.
    * @return A Dictionary that represents the contents of the metadata.
    */