+++ ReleaseTooling/Sources/ZipBuilder/ZipBuilder.swift
@@ -369,11 +369,8 @@ struct ZipBuilder {
     guard !podsToInstall.isEmpty else {
       fatalError("Failed to find versions for Firebase release")
     }
-    // We don't release Google-Mobile-Ads-SDK and GoogleSignIn, but we include their latest
+    // We don't release GoogleSignIn, but we include its latest
     // version for convenience in the Zip and Carthage builds.
-    podsToInstall.append(CocoaPodUtils.VersionedPod(name: "Google-Mobile-Ads-SDK",
-                                                    version: nil,
-                                                    platforms: ["ios"]))
     podsToInstall.append(CocoaPodUtils.VersionedPod(name: "GoogleSignIn",
                                                     version: nil,
                                                     platforms: ["ios"]))
@@ -486,8 +483,7 @@ struct ZipBuilder {
 
     // Skip Analytics and the pods bundled with it.
     let remainingPods = installedPods.filter {
-      $0.key == "Google-Mobile-Ads-SDK" ||
-        $0.key == "GoogleSignIn" ||
+      $0.key == "GoogleSignIn" ||
         (firebaseZipPods.contains($0.key) &&
           $0.key != "FirebaseAnalytics" &&
           $0.key != "Firebase" &&