+++ FirebaseAI/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 - [added] Added support for Grounding with Google Search. (#15014)
+- [removed] Removed `CountTokensResponse.totalBillableCharacters` which was
+  deprecated in 11.15.0. Use `totalTokens` instead. (#15056)
 
 # 11.15.0
 - [fixed] Fixed `Sendable` warnings introduced in the Xcode 26 beta. (#14947)