+++ FirebaseRemoteConfig/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+- [added] Improved how the SDK handles real-time requests when a Firebase
+  project has exceeded its available quota for real-time services.
+  Released in anticipation of future quota enforcement, this change is
+  designed to fetch the latest template even when the quota is exhausted.
+
 # 11.14.0
 - [fixed] Fix build warning from comparison of different enumeration types.